refactor(domain): `UserNotFoundException` 도입 및 관련 코드 리팩토링
- `NoSuchElementException`을 `UserNotFoundException`으로 대체
- jOOQ 관련 유틸 확장 메서드 수정 및 안전성을 위한 `mapNotNull` 도입
- 테스트 코드에 새 예외 클래스 및 메시지 반영
- jOOQ로 생성된 테이블/레코드 클래스 및 관련 환경 파일 추가
- `.gitignore`에 jOOQ 관련 설정 업데이트 및 불필요 파일 방지 처리
- `FindUserQueryHandler` 리팩토링 및 `UserQueryRepository` 인터페이스 활용

+++ common/data/src/main/kotlin/com/nextmall/common/data/jooq/util/JooqPaginationExtensions.kt
@@ -56,7 +56,7 @@ fun <R : Record, T> SelectLimitStep<R>.fetchPage(
     size: Int,
     mapper: (R) -> T
 ): PageResult<T> {
-    val offset = page * size
+    val offset = page.toLong() * size
 
     // COUNT(*) query
     val total = DSL.using(this.configuration())
